check for dangling references in callbacks
Using Python's internal reference counting, we can check if
the callback has kept any references to the OSM object that
was handed in. Using pybind11's the keep_alive() policy this
can even be extended to members of the object like the tag
or node list. If dangling references are detected we bail out
with an exception.

Solves the issue that users got unmotivated segfaults when
keeping references.