tools: introduce a socket timeout
Python has blocking connections per default which means that
the download tools can hang indefinitely under some circumstances.
Introduce a default timeout of 60 seconds. It is possible to
change the value via the command line. If the replication runs
into a timeout it stops the download and salvages what was done
so far.