fix off-by-one issues in pyosmium-get-changes
When a start date is given, then start at the change that contains
the date, not the one after.

When writing out end sequences without downloading data, use the
sequence ID **before** the computed start ID, i.e. the of the previous
run. Note that this might result in negative state IDs.